+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove full transaction data from address transaction output data.
 - Remove limits from address transaction output data.
 - Remove block data from previous output.
+- Remove spender from JSON response when output not spent.
 
 ## 0.2.3
 ### Removed